ImportParams.OptionString and params support for thumbnail
