+++ vips/image.go
@@ -143,10 +143,7 @@ func (i *ImportParams) OptionString() string {
 	if v := i.HeifThumbnail; v.IsSet() {
 		values = append(values, "thumbnail="+boolToStr(v.Get()))
 	}
-	if len(values) == 0 {
-		return ""
-	}
-	return "[" + strings.Join(values, ",") + "]"
+	return strings.Join(values, ",")
 }
 
 func boolToStr(v bool) string {
@@ -421,39 +418,24 @@ func LoadImageFromBuffer(buf []byte, params *ImportParams) (*ImageRef, error) {
 
 // NewThumbnailFromFile loads an image from file and creates a new ImageRef with thumbnail crop
 func NewThumbnailFromFile(file string, width, height int, crop Interesting) (*ImageRef, error) {
-	startupIfNeeded()
-
-	vipsImage, format, err := vipsThumbnailFromFile(file, width, height, crop, SizeBoth, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	ref := newImageRef(vipsImage, format, nil)
-
-	govipsLog("govips", LogLevelDebug, fmt.Sprintf("created imageref %p", ref))
-	return ref, nil
+	return LoadThumbnailFromFile(file, width, height, crop, SizeBoth, nil)
 }
 
 // NewThumbnailFromBuffer loads an image buffer and creates a new Image with thumbnail crop
 func NewThumbnailFromBuffer(buf []byte, width, height int, crop Interesting) (*ImageRef, error) {
-	startupIfNeeded()
-
-	vipsImage, format, err := vipsThumbnailFromBuffer(buf, width, height, crop, SizeBoth, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	ref := newImageRef(vipsImage, format, buf)
-
-	govipsLog("govips", LogLevelDebug, fmt.Sprintf("created imageref %p", ref))
-	return ref, nil
+	return LoadThumbnailFromBuffer(buf, width, height, crop, SizeBoth, nil)
 }
 
 // NewThumbnailWithSizeFromFile loads an image from file and creates a new ImageRef with thumbnail crop and size
 func NewThumbnailWithSizeFromFile(file string, width, height int, crop Interesting, size Size) (*ImageRef, error) {
+	return LoadThumbnailFromFile(file, width, height, crop, size, nil)
+}
+
+// LoadThumbnailFromFile loads an image from file and creates a new ImageRef with thumbnail crop and size
+func LoadThumbnailFromFile(file string, width, height int, crop Interesting, size Size, params *ImportParams) (*ImageRef, error) {
 	startupIfNeeded()
 
-	vipsImage, format, err := vipsThumbnailFromFile(file, width, height, crop, size, nil)
+	vipsImage, format, err := vipsThumbnailFromFile(file, width, height, crop, size, params)
 	if err != nil {
 		return nil, err
 	}
@@ -466,9 +448,14 @@ func NewThumbnailWithSizeFromFile(file string, width, height int, crop Interesti
 
 // NewThumbnailWithSizeFromBuffer loads an image buffer and creates a new Image with thumbnail crop and size
 func NewThumbnailWithSizeFromBuffer(buf []byte, width, height int, crop Interesting, size Size) (*ImageRef, error) {
+	return LoadThumbnailFromBuffer(buf, width, height, crop, size, nil)
+}
+
+// LoadThumbnailFromBuffer loads an image buffer and creates a new Image with thumbnail crop and size
+func LoadThumbnailFromBuffer(buf []byte, width, height int, crop Interesting, size Size, params *ImportParams) (*ImageRef, error) {
 	startupIfNeeded()
 
-	vipsImage, format, err := vipsThumbnailFromBuffer(buf, width, height, crop, size, nil)
+	vipsImage, format, err := vipsThumbnailFromBuffer(buf, width, height, crop, size, params)
 	if err != nil {
 		return nil, err
 	}
