+++ vips/image_golden_test.go
@@ -410,85 +410,6 @@ func TestImage_Thumbnail_CropCentered(t *testing.T) {
 		}, nil)
 }
 
-func TestThumbnail_NoCrop(t *testing.T) {
-	goldenCreateTest(t, resources+"jpg-8bit-grey-icc-dot-gain.jpg",
-		func(path string) (*ImageRef, error) {
-			return NewThumbnailFromFile(path, 36, 36, InterestingNone)
-		},
-		func(buf []byte) (*ImageRef, error) {
-			return NewThumbnailFromBuffer(buf, 36, 36, InterestingNone)
-		},
-		nil,
-		func(result *ImageRef) {
-			assert.Equal(t, 36, result.Width())
-			assert.Equal(t, 24, result.Height())
-			assert.Equal(t, ImageTypeJPEG, result.Format())
-		}, nil)
-}
-
-func TestThumbnail_NoUpscale(t *testing.T) {
-	goldenCreateTest(t, resources+"jpg-8bit-grey-icc-dot-gain.jpg",
-		func(path string) (*ImageRef, error) {
-			return NewThumbnailWithSizeFromFile(path, 9999, 9999, InterestingNone, SizeDown)
-		},
-		func(buf []byte) (*ImageRef, error) {
-			return NewThumbnailWithSizeFromBuffer(buf, 9999, 9999, InterestingNone, SizeDown)
-		},
-		nil,
-		func(result *ImageRef) {
-			assert.Equal(t, 715, result.Width())
-			assert.Equal(t, 483, result.Height())
-			assert.Equal(t, ImageTypeJPEG, result.Format())
-		}, nil)
-}
-
-func TestThumbnail_CropCentered(t *testing.T) {
-	goldenCreateTest(t, resources+"jpg-8bit-grey-icc-dot-gain.jpg",
-		func(path string) (*ImageRef, error) {
-			return NewThumbnailFromFile(path, 25, 25, InterestingCentre)
-		},
-		func(buf []byte) (*ImageRef, error) {
-			return NewThumbnailFromBuffer(buf, 25, 25, InterestingCentre)
-		},
-		nil,
-		func(result *ImageRef) {
-			assert.Equal(t, 25, result.Width())
-			assert.Equal(t, 25, result.Height())
-			assert.Equal(t, ImageTypeJPEG, result.Format())
-		}, nil)
-}
-
-func TestThumbnail_PNG_CropCentered(t *testing.T) {
-	goldenCreateTest(t, resources+"png-24bit.png",
-		func(path string) (*ImageRef, error) {
-			return NewThumbnailFromFile(path, 25, 25, InterestingCentre)
-		},
-		func(buf []byte) (*ImageRef, error) {
-			return NewThumbnailFromBuffer(buf, 25, 25, InterestingCentre)
-		},
-		nil,
-		func(result *ImageRef) {
-			assert.Equal(t, 25, result.Width())
-			assert.Equal(t, 25, result.Height())
-			assert.Equal(t, ImageTypePNG, result.Format())
-		}, nil)
-}
-
-func TestThumbnail_Decode_BMP(t *testing.T) {
-	goldenCreateTest(t, resources+"bmp.bmp",
-		func(path string) (*ImageRef, error) {
-			return NewThumbnailWithSizeFromFile(path, 9999, 9999, InterestingNone, SizeDown)
-		},
-		func(buf []byte) (*ImageRef, error) {
-			return NewThumbnailWithSizeFromBuffer(buf, 9999, 9999, InterestingNone, SizeDown)
-		},
-		nil,
-		func(img *ImageRef) {
-			assert.Equal(t, 164, img.Width())
-			assert.Equal(t, 211, img.Height())
-		}, nil)
-}
-
 func TestImage_ResizeWithVScale(t *testing.T) {
 	goldenTest(t, resources+"jpg-24bit.jpg",
 		func(img *ImageRef) error {
